RoutingTable splitted into LocalRouting and IterativeRouting
